Added MapRange function + tests. Not quite working.
The idea behind this was to have an equivalent to the ParallelFor
function that works with any generic callable. I'm testing this with a
class with data members and an operator() function, but the lambda
capture statement causes the functor to be captured by value, thereby
making the data thread-private (meaning that MapRange has no effect on
the main data). I also tried capturing func by reference, but that
causes it to crash. This remains to be fixed.